Support binary only runner (#21)
Support a possibility to use class path as the source for features, like this: `features = "classpath:features/"`
It means that we can skip uploading feature files and just use those embedded in the binary.
The code is hacky, again.
For whatever reason cucumber doesn't like when classpath path starts with a slash, e.g. `classpath:/feature/` and wants it to be `classpath:feature/`. Maybe that is a java thing, but I figured that I might as well just string replace that.

Unrelated changes:
 - return 0 children instead of an exception. Needed for the JUnit4 runner when started from commandline
 - Add some console logging, it's better than no logging at all
 - badge :D