Add support for multiple annotations for each class. (#30)
The reason behind the decision is that currently we can't run different scenarios within the same project in parallel. One way to do it would be to parallelize on JUnit level, but at this moment JUnit5 (which is the one that supports parallel executions) is not supported. Therefore, we are going with multiple annotations on the same class. This is done by flattening the features from all the annotations into one "Job".

Cons:

Currently threads are set per annotation. So if an override is used it would override both settings. So, if you have 15 + 10 threads set (25 in total) and would override them to say 30, the result would be 30+30=60.